修改首页和视频控件的计时器,改用handler模式,取消timer模式(这个需要后面验证,特别是视频)

+++ app/src/main/java/com/caotu/duanzhi/utils/DataCleanManager.java
@@ -12,6 +12,7 @@ import java.math.BigDecimal;
 public class DataCleanManager {
     /**
      * 获取缓存大小
+     *
      * @param context
      * @return
      * @throws Exception
@@ -26,6 +27,7 @@ public class DataCleanManager {
 
     /**
      * 清除缓存
+     *
      * @param context
      */
     public static void clearAllCache(Context context) {
@@ -44,6 +46,9 @@ public class DataCleanManager {
                     return false;
                 }
             }
+        }
+        if (dir == null) {
+            return false;
         }
         return dir.delete();
     }
@@ -71,6 +76,7 @@ public class DataCleanManager {
 
     /**
      * 格式化单位
+     *
      * @param size
      * @return
      */
